1.优化Action的参数，可为空 2.增加文件上传支持

+++ src/main/java/com/dean4j/framework/uitl/CastUtil.java
@@ -75,7 +75,7 @@ public final class CastUtil {
     /**
      * 转为int型
      */
-    public static long castInt(Object obj) {
+    public static int castInt(Object obj) {
         return CastUtil.castInt(obj, 0);
     }
